Namespace layout components to avoid conflicts
